requirements.txt
fastapi
uvicorn
python-multipart
pymongo
python-dotenv
spacy
pdfminer.six
scikit-learn
